networking whitepaper final

+++ 4.1.9-networking-whitepaper/run_loop.py
@@ -38,11 +38,17 @@ def run_cmd(cmd, run_env=False, unsafe_shell=False, check_rc=False):
             sys.exit(p.returncode)
     return p.returncode
 
-def set_delay(device, milliseconds):
-    run_cmd(['sudo', 'tc', 'qdisc', 'del', 'dev', device, 'root', 'netem'], check_rc=True)
-    run_cmd(['sudo', 'tc', 'qdisc', 'add', 'dev', device, 'root', 'netem', 'delay', '{0}ms'.format(milliseconds)], check_rc=True)
-
-
+def set_delay(connection_string, device, total_milliseconds):
+    each_milliseconds = int(total_milliseconds) / 2.0
+    cmds = [
+        ['sudo', 'tc', 'qdisc', 'del', 'dev', device, 'root', 'netem'],
+        ['sudo', 'tc', 'qdisc', 'add', 'dev', device, 'root', 'netem', 'delay', '{0}ms'.format(each_milliseconds)]
+    ]
+    for cmd in cmds:
+        run_cmd(cmd, check_rc=True)
+    for cmd in cmds:
+        remotecmd = ['ssh', '-t', connection_string, " ".join(cmd)]
+        run_cmd(remotecmd, check_rc=True)
 
 def set_tcp_settings(connection_string, size):
     if size == 'default':
@@ -67,7 +73,6 @@ def set_tcp_settings(connection_string, size):
         remotecmd = ['ssh', '-t', connection_string, " ".join(cmd)]
         run_cmd(remotecmd, check_rc=True)
 
-
 def set_parallel_buffer_size(connection_string, megabytes):
     keyname='irods_transfer_buffer_size_for_parallel_transfer_in_megabytes'
 
@@ -85,16 +90,17 @@ def set_parallel_buffer_size(connection_string, megabytes):
     localcmd.extend([filename])
     run_cmd(localcmd, check_rc=True)
 
-
 def main():
     # check parameters
-    usage = 'Usage: %prog [options] user@host delay_device results_file'
+    usage = 'Usage: %prog [options] iput_or_iget user@host delay_device results_file'
     parser = optparse.OptionParser(usage)
     parser.add_option('-q', '--quiet', action='store_const', const=0, dest='verbosity', help='print less information to stdout')
     parser.add_option('-v', '--verbose', action='count', dest='verbosity', default=1, help='print more information to stdout')
     (options, args) = parser.parse_args()
-    if len(args) != 3:
+    if len(args) != 4:
         parser.error('incorrect number of arguments')
+    if args[0] not in ['iput', 'iget']:
+        parser.error("transfer command [{0}] must be either 'iput' or 'iget'".format(args[0]))
     if len(args) == 0:
         parser.print_usage()
         return 1
@@ -113,9 +119,10 @@ def main():
     log.addHandler(ch)
 
     # setup
-    connection_string = args[0]
-    delay_device = args[1]
-    results_file = args[2]
+    transfer_command = args[0]
+    connection_string = args[1]
+    delay_device = args[2]
+    results_file = args[3]
     targetdir = 'results'
     itargetdir = 'speedtest'
 
@@ -126,36 +133,56 @@ def main():
     num_threads_array = ['1', '2', '3', '4', '8', '16']
     runs_of_each      = 3
 
+    filesize_array    = ['1', '5', '10']
+    delay_array       = ['0']
+    tcp_size_array    = ['big']
+    buffer_size_array = ['100']
+    num_threads_array = ['3', '4']
+    runs_of_each      = 1
+
     # csv headers
     with open(os.path.join(script_path, results_file), 'a') as f:
-        f.write('GB,delay,tcp_size,parallel_buffer,N,run,seconds\n')
+        f.write('command,MiB,delay,tcp_size,parallel_buffer,N,run,seconds\n')
 
     # BIGLOOP
     for filesize in filesize_array:
-        filename = '{0}Gfile'.format(filesize)
-        run_cmd(['truncate', '-s{0}G'.format(filesize), filename], check_rc=True)
-        mkdir_p(os.path.join(script_path, targetdir, '{0}G'.format(filesize)))
+        filename = '{0}Mfile'.format(filesize)
+        run_cmd(['truncate', '-s{0}M'.format(filesize), filename], check_rc=True)
+        if transfer_command == 'iput':
+            mkdir_p(os.path.join(script_path, targetdir, '{0}M'.format(filesize)))
+        if transfer_command == 'iget':
+            run_cmd(['iput', '-f', '-N3', filename])
+            os.unlink(filename)
         for delay in delay_array:
-            set_delay(delay_device, delay)
+            set_delay(connection_string, delay_device, delay)
             for tcp_size in tcp_size_array:
                 set_tcp_settings(connection_string, tcp_size)
                 for buffer_size in buffer_size_array:
                     set_parallel_buffer_size(connection_string, buffer_size)
                     for num_threads in num_threads_array:
                         for run in range(1, runs_of_each+1):
-                            run_cmd(['imkdir', '-p', itargetdir], check_rc=True)
-                            ifilename = '{0}/N{1}'.format(itargetdir, num_threads)
-                            cmd = ['iput', '-v', '-N{0}'.format(num_threads), filename, ifilename]
+                            if transfer_command == 'iput':
+                                run_cmd(['imkdir', '-p', itargetdir], check_rc=True)
+                                ifilename = '{0}/N{1}'.format(itargetdir, num_threads)
+                                cmd = ['iput', '-v', '-N{0}'.format(num_threads), filename, ifilename]
+                            if transfer_command == 'iget':
+                                cmd = ['iget', '-v', '-N{0}'.format(num_threads), filename]
                             start = time.time()
                             run_cmd(cmd, check_rc=True)
                             end = time.time()
                             duration = end - start
                             with open(os.path.join(script_path, results_file), 'a') as f:
-                                line = [filesize, delay, tcp_size, buffer_size, num_threads, str(run), str(duration)]
+                                line = [transfer_command, filesize, delay, tcp_size, buffer_size, num_threads, str(run), str(duration)]
                                 f.write(','.join(line))
                                 f.write('\n')
-                            run_cmd(['irm', '-rf', itargetdir], check_rc=True)
-        os.unlink(filename)
+                            if transfer_command == 'iput':
+                                run_cmd(['irm', '-rf', itargetdir], check_rc=True)
+                            if transfer_command == 'iget':
+                                os.unlink(filename)
+        if transfer_command == 'iput':
+            os.unlink(filename)
+        if transfer_command == 'iget':
+            run_cmd(['irm', '-f', filename], check_rc=True)
 
 
 if __name__ == '__main__':
